+++ sdk/ai/azure-ai-agents/src/test/java/com/azure/ai/agents/models/WebSearchApproximateLocationSerializationTests.java
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.ai.agents;
+package com.azure.ai.agents.models;
 
-import com.azure.ai.agents.models.WebSearchApproximateLocation;
 import com.azure.json.JsonProviders;
 import com.azure.json.JsonReader;
 import com.azure.json.JsonWriter;
